Fix broken unittests
